FTS bug fix

+++ org.zenframework.z8.server/src/main/java/org/zenframework/z8/server/db/sql/fts/FtsConfig.java
@@ -46,7 +46,7 @@ public class FtsConfig extends OBJECT {
 	public FtsQueryType queryType = FtsQueryType.Plain;
 	@SuppressWarnings("rawtypes")
 	public RCollection weight = null;
-	public integer normalization = new integer(10);
+	public integer normalization = new integer(0);
 	public bool coatingDensity = bool.False;
 
 }
